refactor(steam): extract SteamVdfParser and SteamArtworkService from SteamService
SteamVdfParser: all VDF/ID parsing methods (ParseLibraryFolders, ParseAppManifest,
ParseInstallDir, ParseShortcuts, ParseExeField, IsShortcutAppId, GenerateShortcutAppId)

SteamArtworkService: artwork resolution (FindArtworkPath, GetArtworkDiagnostics)

SteamService now delegates to these for parsing and artwork, keeping only
game management, running detection, launching, stopping, and bindings.
